memblock tests: add memblock_alloc_raw tests
Add tests for memblock_alloc_raw() for both bottom_up and top_down
allocation directions. The tests are very similar to the existing tests
for memblock_alloc(), except they do not check if the memory is zeroed.

The tested scenarios are:
- Region can be allocated on the first fit (with and without
  region merging)
- Region can be allocated on the second fit (with and without
  region merging)
- Region can be allocated between two already existing entries
- Limited memory available
- Region is larger than the total size of available memory
- All memory is reserved
- No available memory registered with memblock

Add test case wrappers to test both directions in the same context.

Signed-off-by: Rebecca Mckeever <remckee0@gmail.com>