memblock tests: add memblock_alloc_exact_nid_raw tests
Add tests for memblock_alloc_exact_nid_raw().

The tests are very similar to the existing tests for
memblock_alloc_try_nid. Both top-down and bottom-up allocation
directions are tested. The tested scenarios are:

The tested scenarios are:
- Region can be allocated within the requested range (both with aligned
and misaligned boundaries)
- Region can be allocated between two already existing entries
- Not enough space between already reserved regions
- Memory at the range boundaries is reserved but there is enough space
to allocate a new region
- Memory range is too narrow but memory can be allocated before
the maximum address
- Edge cases:
+ Minimum address is below memblock_start_of_DRAM()
+ Maximum address is above memblock_end_of_DRAM()
- Region starts at the min_addr and ends at max_addr
- Maximum address is too close to the beginning of the available
memory

Add test case wrappers to test both directions in the same context.

Signed-off-by: Rebecca Mckeever <remckee0@gmail.com>