dev: x86/numa: rework numa_clear_kernel_node_hotplug to use memblock
numa_clear_kernel_node_hotplug() marks all modes containing reserved
memory as not-hotswappable.

For that add setting of the node on the reserved ranges at
numa_add_memblk() time and then use the information in memblock to
ensure that nodes that have reserved memory (primarily kernel image and
firmware areas) are marked as not hotswappable.

Signed-off-by: Mike Rapoport <rppt@linux.ibm.com>